Add support for Nemotron Omni (#46509)
* Add Nemotron Omni model

* Declare patch_size, img_context_token_id, video_context_token_id, and related token fields explicitly in NemotronH_Nano_Omni_Reasoning_V3_Config

* Use language model from models/nemotron_h

* Remove processing script

* Move audio model into same model file

* Change RMSNorm to fp32

* Fix rms import

* Update Radio config and model

* Add Nemotron Omni in AutoConfig

* Remove Nano in class

* Add doc and tests

* Fix config check

* Fix for consistency pipeline

* Change config init instead of add the allow list

* Fix consistency

* Isolate pixel shuffle

* Add prefix for some module

* Update SoundConfig and SoundEncoder SoundProjector

* Isolate visual feature extract

* Change MLP and SoundEncoder

* Make video_patch_projection as args in Radio instead of manually patching it in omni model

* Change multimodal path

* Update to use parent's generate path

* Update doc

* Reuse MLP for soundprojection

* Remove unused variable

* Fix consistency check

* Remove NemotronH_Omni_Reasoning_V3SoundEncoder wrap

* Remove vision_model in VisionProjector forward args

* Add integration test

* Update for consistency

* Align get_image_features get_video_features with tests

* Remove config config-derived parameters from checkpoint

* Remove legacy config

* Remove helper

* Fix for code quality test

* Add final layernorm for multi token prediction

* Update doc date

* conditional video patch in radio

* Route NemotronH-Omni sub-configs through CONFIG_MAPPING/AutoConfig

* Keep parameter for campability

* Update config file

* Update processor

* Update nemotron_h_omni modeling file

* Update radio

* Update tests

* Add nemotron_h changes

* last nits to merge!

---------

+++ src/transformers/models/nemotron_h/modeling_nemotron_h.py
@@ -596,7 +596,9 @@ class NemotronHRMSNorm(nn.Module):
         hidden_states = hidden_states.to(torch.float32)
         variance = hidden_states.pow(2).mean(-1, keepdim=True)
         hidden_states = hidden_states * torch.rsqrt(variance + self.variance_epsilon)
-        return self.weight * hidden_states.to(input_dtype)
+        # Unlike Llama, the weight multiply is kept in fp32 and only the result is cast back to the input
+        # dtype, matching the reference implementation.
+        return (self.weight.to(torch.float32) * hidden_states).to(input_dtype)
 
     def extra_repr(self):
         return f"{tuple(self.weight.shape)}, eps={self.variance_epsilon}"
